refactor/changed routing
