+++ modules/quiz/quiz-postrequest.js
@@ -23,45 +23,38 @@ module.exports = app => {
     sportQuizData.map(x => {
       let percentage = 0
 
-      const groupCalc = (() => {
-        const groupOrSoloInObject =
-          x.groupOrSolo === yourResultsOfForm.groupOrSolo
-        if (groupOrSoloInObject === true) {
-          percentage += 33.33
-        }
-        if (groupOrSoloInObject === false) {
-          const calcedPercentage =
-            33.33 -
-            Math.abs(yourResultsOfForm.groupOrSolo - x.groupOrSolo) * 3.33
-          percentage += calcedPercentage
-        }
-      })()
+      const groupOrSoloInObject =
+        x.groupOrSolo === yourResultsOfForm.groupOrSolo
+      if (groupOrSoloInObject === true) {
+        percentage += 33.33
+      }
+      if (groupOrSoloInObject === false) {
+        const calcedPercentage =
+          33.33 - Math.abs(yourResultsOfForm.groupOrSolo - x.groupOrSolo) * 3.33
+        percentage += calcedPercentage
+      }
 
-      const inOutdoorCalc = (() => {
-        const inOrOutdoorInObject =
-          x.inOrOutdoor === yourResultsOfForm.inOrOutdoor
-        if (inOrOutdoorInObject === true) {
-          percentage += 33.33
-        }
-        if (inOrOutdoorInObject === false) {
-          const calcedPercentage =
-            33.33 -
-            Math.abs(yourResultsOfForm.inOrOutdoor - x.inOrOutdoor) * 3.33
-          percentage += calcedPercentage
-        }
-      })()
+      const inOrOutdoorInObject =
+        x.inOrOutdoor === yourResultsOfForm.inOrOutdoor
+      if (inOrOutdoorInObject === true) {
+        percentage += 33.33
+      }
+      if (inOrOutdoorInObject === false) {
+        const calcedPercentage =
+          33.33 - Math.abs(yourResultsOfForm.inOrOutdoor - x.inOrOutdoor) * 3.33
+        percentage += calcedPercentage
+      }
 
-      const fishOrLandCalc = (() => {
-        const fishOrLandInObject = x.fishOrLand === yourResultsOfForm.fishOrLand
-        if (fishOrLandInObject === true) {
-          percentage += 33.33
-        }
-        if (fishOrLandInObject === false) {
-          const calcedPercentage =
-            33.33 - Math.abs(yourResultsOfForm.fishOrLand - x.fishOrLand) * 6.66
-          percentage += calcedPercentage
-        }
-      })()
+      const fishOrLandInObject = x.fishOrLand === yourResultsOfForm.fishOrLand
+      if (fishOrLandInObject === true) {
+        percentage += 33.33
+      }
+      if (fishOrLandInObject === false) {
+        const calcedPercentage =
+          33.33 - Math.abs(yourResultsOfForm.fishOrLand - x.fishOrLand) * 6.66
+        percentage += calcedPercentage
+      }
+
       comparedResultsOfSports.push({
         percentageNum: percentage,
         index: mapCounter
