+++ modules/routes.js
@@ -1,14 +1,11 @@
-let allSports;
-
-(async function () {
-  const getAllSports = require("./sportslist.js")
+let allSports
+;(async function() {
+  const getAllSports = require("./sportlist/sportlist-az-list.js")
   allSports = await getAllSports()
 })()
 
-const getEvents = require("./sportslist-events.js")
-
-const quizPostRequest = require("./quizPostRequest.js"),
-  fetchData = require("./fetch.js")
+const quizPostRequest = require("./quiz/quiz-postrequest.js"),
+  fetchData = require("./helper/helper-fetch.js")
 
 module.exports = (app, eventsData) => {
   app.get("/", (req, res) => {
@@ -18,7 +15,7 @@ module.exports = (app, eventsData) => {
     })
   })
   app.get("/sportslist", async (req, res) => {
-    res.render("pages/sportslist.ejs", {
+    res.render("pages/sportlist/sportlist-az-list.ejs", {
       hero: "small-hero",
       heroText: ["Sports Activities A-Z"],
       data: allSports,
@@ -26,14 +23,14 @@ module.exports = (app, eventsData) => {
     })
   })
   app.get("/sportsmap", async (req, res) => {
-    res.render("pages/sportsmap.ejs", {
+    res.render("pages/sportlist/sportlist-map.ejs", {
       hero: "small-hero",
       heroText: ["Sports Activities A-Z MAP"]
     })
   })
   app.get("/sportslist/clubs/:id", async (req, res) => {
     const id = req.params.id
-    res.render("pages/sportlist-clubs.ejs", {
+    res.render("pages/sportlist/sportlist-clubs.ejs", {
       hero: "small-hero",
       heroText: ["Sports Activities A-Z"],
       sport: id
@@ -41,44 +38,41 @@ module.exports = (app, eventsData) => {
   })
   app.get("/sportslist/events/:id", async (req, res) => {
     const id = req.params.id
-    const filteredEvents = await getEvents(id)
-    console.log(filteredEvents)
-    res.render("pages/sportlist-events.ejs", {
+    res.render("pages/sportlist/sportlist-events.ejs", {
       hero: "small-hero",
       heroText: ["Sports Activities A-Z"],
-      events: filteredEvents,
       sport: id
     })
   })
   app.get("/events", (req, res) => {
-    res.render("pages/events.ejs", {
+    res.render("pages/events/events-overview.ejs", {
       hero: "small-hero",
       heroText: ["Events"],
       data: eventsData
     })
   })
   app.get("/quiz", (req, res) => {
-    res.render("pages/quiz.ejs", {
+    res.render("pages/quiz/quiz-questions.ejs", {
       hero: "small-hero",
       heroText: ["Amsterdam", "Zuid-Oost", "Be a part of it!"]
     })
   })
   app.get("/login", (req, res) => {
-    res.render("pages/login.ejs", {
+    res.render("pages/sportprovider/sportprovider-login.ejs", {
       hero: "small-hero",
       heroText: ["Login/Register"]
     })
   })
   app.get("/sportprovider", (req, res) => {
-    res.render("pages/sportprovider-dashboard.ejs", {
+    res.render("pages/sportprovider/sportprovider-dashboard.ejs", {
       hero: "small-hero",
       heroText: ["Dashboard"]
     })
   })
-  app.get("/create-event", async (req, res) => {
+  app.get("/sportprovider/sportprovider-create-event", async (req, res) => {
     const data = await fetchData()
 
-    res.render("pages/create-event.ejs", {
+    res.render("pages/sportprovider/sportprovider-create-event.ejs", {
       hero: "small-hero",
       heroText: ["Create Event"],
       data: data
